// src/app/admin/inventaires/shared/model/arbre.interface.ts
import { Espece } from "../../../config-inventaire/shared/model/espece.interface";
import { Champignons } from "../../../config-inventaire/shared/model/Champignons";
import { Coordinate, ImgInventaire } from "./inventaire.interface";
import { Nuisible } from "../../../config-inventaire/shared/model/nuisible.interface";
import { Resource } from "../../../../shared/modele/resource";

export interface Arbre extends Resource {
  espece: Espece;
  diametre: Number;
  hauteur: Number;
  coord: Coordinate;
  imgUrl?: ImgInventaire;
  codeSite?: String;
  numSujet?: String;
  critere?: String[];
  implantation?: String;
  domaine?: String;
  nuisance?: String[];
  proximite?: String[];
  proximiteWithDict?: String[];
  tauxFreq?: String;
  typePassage?: String[];
  accessibilite?: String;
  abattage?: String;
  travauxColletOther?: String;
  travauxTroncOther?: String;
  travauxHouppierOther?: String;
  dateTravaux?: String;
  userEditedDateTravaux: any;
  dateProVisite?: String;
  comProVisite?: String;
  caractPied?: String[];
  caractTronc?: String;
  portArbre?: String;
  stadeDev?: String;
  etatSanCollet?: String[];
  etatSanTronc?: String[];
  etatSanHouppier?: String[];
  nuisanceNuisibles?: Nuisible[];
  address?: String;
  dict?: String[];
  risque?: any;
  etatSanGeneral?: String[];
  etatSanGeneralOther?: String;
  proximiteOther?: String;
  caractPiedOther?: String;
  accessibiliteOther?: String;
  caractTroncMultiples?: Number;
  etatSanColletCavite?: Number;
  etatSanColletChampignons?: Champignons[];
  etatSanTroncCavite?: Number;
  etatSanTroncCorpsEtranger?: String;
  etatSanTroncChampignons?: Champignons[];
  etatSanTroncNuisibles?: Nuisible[];
  etatSanHouppierChampignons?: Champignons[];
  etatSanHouppierNuisibles?: Nuisible[];
  risqueGeneral?: String[];
  travauxTroncProtection?: String;
  critereOther?: String;
  risqueGeneralOther?: String;
  typePassageOther?: String;
  typeIntervention?: String;
  statusTravaux: boolean;
  varietyGrade?: any;
  healthIndex?: any;
  aestheticIndex?: any;
  locationIndex?: any;
  aestheticColumn: any;
  healthColumn: any;
  ville: String;
  pays: String;
  beva?: any;
  travauxColletMultiple?: string[];
  travauxTroncMultiple?: string[];
  travauxHouppierMultiple?: string[];
  etatSanTroncNuisiblesAutres?: string;
  etatSanHouppierNuisiblesAutres?: string;
  etatSanColletChampignonsAutres?: string;
  etatSanTroncChampignonsAutres?: string;
  etatSanHouppierChampignonsAutres?: string;
  // etat san other
  etatSanColletOther?: string;
  etatSanTroncOther?: string;
  etatSanHouppierOther?: string;
  type?: string;
}